+++ HISTORY.rst
@@ -23,6 +23,10 @@ latest (unreleased)
 
 **Bugfixes**
 
+* BSMTS-279 : Uninstall DJ modules
+* BSMTS-279 : Update XMLIDs of CSV imported taxes, tags, and report lines with wrong module name
+* BSMTS-279 : Recreate BI View Engagements
+
 **Build**
 
 * Update Dockerimage to 10.0-2.4.1